invoice-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4
anthropic==0.7.8
openai==1.12.0
# boto3 removed - using local storage instead of S3
pandas>=2.1.4
openpyxl>=3.1.2
python-dotenv>=1.0.0
httpx>=0.25.2
pydantic[email]>=2.5.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
faker>=20.1.0
structlog>=23.2.0
rich>=13.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
tenacity==8.2.3
numpy>=1.24.4
agentic-doc==0.1.0
